+++ installer/db/mysql/mediadb/init_01_create-database.sql
@@ -1,5 +1,5 @@
 -- ------------------------------------
 -- create current Database: `mediadb`
 -- ------------------------------------
-CREATE DATABASE IF NOT EXISTS mediadb DEFAULT CHARACTER SET latin1 COLLATE latin1_general_ci;
-USE `mediadb`;
+CREATE DATABASE IF NOT EXISTS testmediadb DEFAULT CHARACTER SET latin1 COLLATE latin1_general_ci;
+USE `testmediadb`;