+++ installer/db/mysql/mediadb/init_02_create-user.sql
@@ -1,13 +1,13 @@
 -- ------------------------------------
 -- create users
 -- ------------------------------------
-CREATE USER IF NOT EXISTS 'mediadb@127.0.0.1' IDENTIFIED BY 'mediadb';
-GRANT ALL PRIVILEGES ON mediadb.* TO 'mediadb@127.0.0.1' IDENTIFIED BY 'mediadb';
-CREATE USER IF NOT EXISTS 'mediadb@localhost' IDENTIFIED BY 'mediadb';
-GRANT ALL PRIVILEGES ON mediadb.* TO 'mediadb@localhost' IDENTIFIED BY 'mediadb';
-CREATE USER IF NOT EXISTS 'mediadb@*' IDENTIFIED BY 'mediadb';
-GRANT ALL PRIVILEGES ON mediadb.* TO 'mediadb@*' IDENTIFIED BY 'mediadb';
-CREATE USER IF NOT EXISTS 'mediadb' IDENTIFIED BY 'mediadb';
-GRANT ALL PRIVILEGES ON mediadb.* TO 'mediadb' IDENTIFIED BY 'mediadb';
+CREATE USER IF NOT EXISTS 'testmediadb@127.0.0.1' IDENTIFIED BY 'testmediadb';
+GRANT ALL PRIVILEGES ON testmediadb.* TO 'testmediadb@127.0.0.1' IDENTIFIED BY 'testmediadb';
+CREATE USER IF NOT EXISTS 'testmediadb@localhost' IDENTIFIED BY 'testmediadb';
+GRANT ALL PRIVILEGES ON testmediadb.* TO 'testmediadb@localhost' IDENTIFIED BY 'testmediadb';
+CREATE USER IF NOT EXISTS 'testmediadb@*' IDENTIFIED BY 'testmediadb';
+GRANT ALL PRIVILEGES ON testmediadb.* TO 'testmediadb@*' IDENTIFIED BY 'testmediadb';
+CREATE USER IF NOT EXISTS 'testmediadb' IDENTIFIED BY 'testmediadb';
+GRANT ALL PRIVILEGES ON testmediadb.* TO 'testmediadb' IDENTIFIED BY 'testmediadb';
 
 FLUSH PRIVILEGES;