+++ installer/db/mysql/mytb/init_01_create-database.sql
@@ -1,6 +1,0 @@
--- ------------------------------------
--- create current Database: `mytb`
--- ------------------------------------
-CREATE DATABASE IF NOT EXISTS mytb DEFAULT CHARACTER SET latin1 COLLATE latin1_general_ci;
-
-USE `mytb`;